refactor: separate main menu modules
